requirements.txt
altair==5.3.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
gitdb==4.0.11
GitPython==3.1.43
idna==3.6
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.3.0
protobuf==4.25.3
pyarrow==15.0.2
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
streamlit==1.33.0
tenacity==8.2.3
regex==2023.12.25
tiktoken==0.6.0
tree-sitter==0.21.3
tree_sitter_languages
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
setuptools==69.2.0
# FastAPI backend dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
# Search and ML dependencies
opensearch-py==2.4.2
sentence-transformers==2.2.2
# RAG Filter-Rerank dependencies
python-dotenv==1.0.0
pydantic-settings==2.1.0
diskcache==5.6.3
redis==5.0.1
pytest==7.4.3
pytest-mock==3.12.0
